exit status and arrangements
