+++ builtin_exit.c
@@ -6,8 +6,7 @@
  * 
  * Return: 0 when success.
  */
-int builtin_exit(__attribute__((unused)) shdata_t *data)
+int builtin_exit(shdata_t *data)
 {
-	printf("im exit\n");
-	return (0);
+	_exit(data->status);
 }
