+++ main.h
@@ -5,12 +5,16 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
+#include <sys/wait.h>
+#include <sys/stat.h>
 
+extern char **environ;
 
 typedef struct shdata_s
 {
 	char **argv;
 	char **sh_env;
+	int status;
 	int pid;
 } shdata_t;
 
@@ -29,6 +33,7 @@ typedef struct builtin_func_s
 /* string functions */
 int _strcmp(char *, char *);
 char **split_string(const char *str, const char *delimiter);
+int _atoi(char *s);
 
 /* built-in functions */
 int (*get_builtin(char *command))(shdata_t *);
@@ -44,5 +49,6 @@ list_t *add_node_end(list_t **, char *);
 void run_shell(shdata_t *);
 void non_interactive(char **argv);
 int fill_shdata(shdata_t *);
+
 
 #endif
