+++ main_shell.c
@@ -1,7 +1,4 @@
 #include "main.h"
-#include <unistd.h>
-#include <sys/wait.h>
-#include <sys/stat.h>
 
 extern char **environ;
 
@@ -10,12 +7,14 @@ extern char **environ;
  * @data: the empty structure.
  * @argv: main arguments.
  *
- * Return: a filled shdata structure.
+ * Return: 0 when succes and 1 otherwise.
  */
 int fill_shdata(shdata_t *data)
 {
+	/* yes, it is incomplete */
 	int i;
 
+	data->status = 0;
 	for (i = 0; environ[i]; i++)
 		;
 
@@ -32,45 +31,6 @@ int fill_shdata(shdata_t *data)
 	data->sh_env[i] = NULL;
 
 	return (0);
-}
-
-/**
- * non_interactive : runs the shell in non-interactive mode.
- * argv: arguments vector.
- *
- * Return: void.
- */
-
-void non_interactive(char **argv)
-{
-	char *exe_path;
-	int child_state;
-	pid_t child_pid;
-
-	argv++;
-	exe_path = check_exe(argv[0]);
-
-	if (!exe_path)
-	{
-		perror("");
-		exit(1);
-	}
-
-	child_pid = fork();
-	if (child_pid == -1)
-	{
-		perror("");
-		exit(1);
-	}
-
-	if (child_pid == 0)
-	{
-		execve(exe_path, argv, environ);
-		perror("");
-		_exit(1);
-	}
-	else
-		wait(&child_state);
 }
 
 /**
@@ -93,63 +53,4 @@ int main(__attribute__((unused)) int argc, char **argv)
 
 	run_shell(&data);
 	return (0);
-}
-
-/**
- * run_shell - runs the prompt loop.
- * shdata: data of the shell.
- *
- * Return: void.
- */
-void run_shell(shdata_t *data)
-{
-	char *lineptr = NULL, *exe_path;
-	size_t n = 0;
-	int status = 1, child_state;
-	pid_t child_pid;
-	char **args;
-	int (*func)(shdata_t *);
-
-	signal(SIGINT, SIG_IGN);
-
-	while (1)
-	{
-		write(STDOUT_FILENO, "\\(*_*)/ ", 9);
-		fflush(stdout);
-		status = getline(&lineptr, &n, stdin);
-		if (status == -1)
-			break;
-		args = split_string(lineptr, "\n ");
-		func = get_builtin(lineptr);
-		if (func != NULL)
-		{
-			func(data);
-			exit(0);
-		}
-		exe_path = check_exe(args[0]);
-
-		if (!exe_path)
-		{
-			perror("");
-			continue;
-		}
-
-		child_pid = fork();
-		if (child_pid == -1)
-		{
-			perror("");
-			exit(1);
-		}
-
-		if (child_pid == 0)
-		{
-			execve(exe_path, args, data->sh_env);
-			perror("");
-			_exit(1);
-		}
-		else
-			wait(&child_state);
-	}
-
-	free(lineptr);
 }
